Move the AI namespace to Mods.Common
