+++ OpenRA.Mods.Common/AI/AttackOrFleeFuzzy.cs
@@ -14,10 +14,9 @@ using System.Linq;
 using AI.Fuzzy.Library;
 using OpenRA.GameRules;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.AI
+namespace OpenRA.Mods.Common.AI
 {
 	class AttackOrFleeFuzzy
 	{
