+++ OpenRA.Mods.Common/AI/States/StateBase.cs
@@ -13,11 +13,9 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.AI
+namespace OpenRA.Mods.Common.AI
 {
 	abstract class StateBase
 	{
